Remove MessengerTransportImpl
